==> Part II/01_Geojson.ipynb <==
## GEOJSON ##

# import libraries

import pandas as pd
import geopandas as gpd

# import geojson files

geography = gpd.read_file('combined_geographies_geojson.geojson') # output from CP-part1

# to split the multipolygon into individual polygons | results to more rows

geography_clean = geography.explode()
geography_clean = geography_clean.reset_index(drop = True)
geography_clean.head()

# to get the minimun and maximum values of longitude and latitude from the geometry column

geojson_bound_coordinates = geography_clean['geometry']
min_max_xy = geojson_bound_coordinates.bounds

min_max_xy

# to append the min and max values of longitude and latitude to the clean df

geography_clean['min_longitude'] = min_max_xy['minx']
geography_clean['min_latitude'] = min_max_xy['miny']
geography_clean['max_longitude'] = min_max_xy['maxx']
geography_clean['max_latitude'] = min_max_xy['maxy']

# to check updated dataset
geography_clean

# to drop "geometry" column 

geography_clean.drop(columns = ['geometry'],
                             inplace = True)
geography_clean

# to get the min and max values from the duplicated rows

geography_clean = geography_clean.groupby(['GeoUID', 
                                            'Region.Name', 
                                            'provincename']).agg(min_longitude = ('min_longitude', 'min'),
                                                                 max_longitude = ('max_longitude', 'max'),
                                                                 min_latitude = ('min_latitude', 'min'),
                                                                 max_latitude = ('max_latitude', 'max')).reset_index()

# to check data

geography_clean

# to output the file into a csv format

csv_path = 'geography_clean.csv'
geography_clean.to_csv(csv_path, 
                        index = False)
