==> Part II/02_Weather_1998 to 2018-02.ipynb <==
## WEATHER (1998 - 2018/02) ##

# import libraries

import pandas as pd

# load weather data

weather = pd.read_csv('weather_data_clean.csv') # output from CP-part 1
weather

# copied the weather data to weather_clean to keep original data
weather_clean = weather.copy()

# to specify which column and values to drop values from
column_to_check = 'Station Name'
values_to_drop = ['COBOURG (AUT)', 'LANGARA ISLAND RCS', 'PORT WELLER (AUT)',
                  'SANDHEADS CS', 'SATURNA ISLAND CS', 'GILLAM', 'KINCARDINE',
                  'PETAWAWA A']
    # removing these station names because there are not in the geography dataset; for consistency

# to drop columns
weather_clean = weather[~weather[column_to_check].isin(values_to_drop)]
weather_clean.shape

# to replace climate ID to be consistent with the updated climate ID from the Canada Weather Website

updated_climate_ID = {'EDMONTON STONY PLAIN': '301222F',
                    'ATHABASCA 1': '3060L20',
                    'WAINWRIGHT CFB AIRFIELD 21': '301S001',
                    'KITIMAT HATCHERY': '106D289',
                    'SQUAMISH AIRPORT': '10476F0', 
                    'TORONTO NORTH YORK': '615S001',
                    'PORT COLBORNE (AUT)': '613F606',
                    'QUEBEC/JEAN LESAGE INTL': '701S001'}

weather_clean['Climate ID'] = weather_clean['Station Name'].map(lambda x: updated_climate_ID.get(x, None)).fillna(weather_clean['Climate ID'])

# rename weather data column for easier transformation/cleaning
weather_clean.rename(columns = {'Longitude (x)': 'Longitude', 'Latitude (y)': 'Latitude'}, inplace= True)
weather_clean

# to load clean geaography data

geography_clean = pd.read_csv('geography_clean.csv')
geography_clean

# to create a function to check if a point is within the given range
def is_point_in_range(point, min_val, max_val):
    return (point >= min_val) & (point <= max_val)

# to iterate over rows in the weather DataFrame and assign GeoUID based on the conditions
for index, row in weather_clean.iterrows():
    matching_row = geography_clean[
        is_point_in_range(row['Longitude'], geography_clean['min_longitude'], geography_clean['max_longitude']) &
        is_point_in_range(row['Latitude'], geography_clean['min_latitude'], geography_clean['max_latitude'])
    ]
    if not matching_row.empty:
        weather_clean.at[index, 'GeoUID'] = matching_row['GeoUID'].values[0]

# to merge the weather and geography_ontario DataFrames based on 'GeoUID'
merged_data = pd.merge(weather_clean, geography_clean, on='GeoUID', how = 'left')

# to check the count of rows of the original data and the clean data
print(weather_clean.shape)
print(merged_data.shape)

# to check merged data

merged_data.head().T

# to update province name; incorrectly tagged

updated_provincename = {'OTTAWA CDA': 'Ontario',
                    'CHALK RIVER AECL': 'Ontario'}

merged_data['provincename'] = merged_data['Station Name'].map(lambda x: updated_provincename.get(x, None)).fillna(merged_data['provincename'])

# to drop unnecessary columns

merged_data.drop(columns = ['min_longitude',
                            'max_longitude',
                            'min_latitude',
                            'max_latitude'],
                 inplace = True)

# to convert the data type from integer to string

merged_data['GeoUID'] = merged_data['GeoUID'].astype('int').astype('object')
merged_data.info()

# to remove quotation marks from column name: Dir of Max Gust (10's deg)
merged_data.columns = merged_data.columns.str.replace("'", "")

# to output the file into a csv format
csv_path = 'weather_geography_merged_1998_to_2018Feb.csv'
merged_data.to_csv(csv_path, 
                   index = False)

# to extract the station name and climate ID for each provincename; for further weather data extraction

unique_stations = merged_data[['provincename', 'Station Name', 'Climate ID']].drop_duplicates()
unique_stations_df = pd.DataFrame(unique_stations, columns = ['provincename', 'Station Name', 'Climate ID'])

csv_path = 'unique_stations_df.csv'
unique_stations_df.to_csv(csv_path,
                            index = False)

# to check test_cleaning output
ontario = merged_data[merged_data['provincename'] == 'Ontario']
ontario = ontario[ontario['Date/Time'] == '2017-12']
ontario.head().T


checking = merged_data[merged_data['Station Name'] == 'BARRIERE TOURILLI']
checking

checking = merged_data[merged_data['Climate ID'] == '7010480']
checking

checking = merged_data[merged_data['Station Name'] == 'OTTAWA CDA']
checking

check = merged_data[(merged_data['Station Name'] == 'ENTRANCE ISLAND') & (merged_data['Date/Time'] == '1998-01-01')]
check
