==> Part II/03_Weather_2007 to 2017_SK.ipynb <==
## WEATHER (2007 - 2017) FOR SASKATCHEWAN

import pandas as pd
import os

# ignore all warnings

import warnings
warnings.filterwarnings("ignore")

# Directory containing your CSV files
directory = "C:/Users/patri/Downloads/Data Analytics for Business/Winter 24/DAB422-Capstone Project II/Capstone Project/Weather_2007 to 2017_SK"
# TO CHANGE IF RUN BY DIFFERENT USER

# List all CSV files in the directory
files = [file for file in os.listdir(directory) if file.startswith('en_climate_daily_') and file.endswith('.csv')]

# Initialize an empty DataFrame
combined_csv = pd.DataFrame()

# Read and concatenate CSV files
for file in files:
    filepath = os.path.join(directory, file)
    csv = pd.read_csv(filepath)
    combined_csv = pd.concat([combined_csv, csv], ignore_index=True)

# Display the resulting combined DataFrame
print(combined_csv)


# copied the weather data to weather_clean to keep original data
weather_clean = combined_csv.copy()

# checking 'Climate ID' 1047FF0 is not included to be consistence with historical weather data

sk_climate_id = weather_clean[weather_clean['Climate ID'] == "1047FF0"]

sk_climate_id

# to specify which column and values to drop values from
column_to_check = 'Station Name'
values_to_drop = ['COBOURG (AUT)', 'LANGARA ISLAND RCS', 'PORT WELLER (AUT)',
                  'SANDHEADS CS', 'SATURNA ISLAND CS', 'GILLAM', 'KINCARDINE']
    # removing these station names because there are not in the geography dataset; for consistency | checked using Excel
    # GILLAM is for Manitoba
    # KINCARDINE, data is only until 2022; removed for data range consistency
    
# to drop columns
weather_clean = weather_clean[~weather_clean[column_to_check].isin(values_to_drop)]
weather_clean.shape

# rename column names for easier transformation/cleaning
weather_clean.rename(columns = {'Longitude (x)': 'Longitude', 'Latitude (y)': 'Latitude'}, inplace=True)

# to load clean geaography data
geography_clean = pd.read_csv('geography_clean.csv')
geography_clean

# to create a function to check if a point is within the given range
def is_point_in_range(point, min_val, max_val):
    return (point >= min_val) & (point <= max_val)

# to iterate over rows in the weather DataFrame and assign GeoUID based on the conditions
for index, row in weather_clean.iterrows():
    matching_row = geography_clean[
        is_point_in_range(row['Longitude'], geography_clean['min_longitude'], geography_clean['max_longitude']) &
        is_point_in_range(row['Latitude'], geography_clean['min_latitude'], geography_clean['max_latitude'])
    ]
    if not matching_row.empty:
        weather_clean.at[index, 'GeoUID'] = matching_row['GeoUID'].values[0]
        
# to merge the weather and geography DataFrames based on 'GeoUID'
merged_data = pd.merge(weather_clean, geography_clean, on='GeoUID', how = 'left')

# to check the count of rows of the original data and the clean data
print(weather_clean.shape)
print(merged_data.shape)

# copying the original merged data to not overwrite the original
monthly_data = merged_data.copy()

# to get mean, and sum of columns
monthly_data_clean = monthly_data.groupby(['Longitude', 
                                      'Latitude',
                                      'Station Name',
                                      'Climate ID',
                                      'Year',
                                      'Month',
                                      'GeoUID',
                                      'Region.Name',
                                      'provincename']).agg({'Max Temp (°C)': 'mean', 
                                                            'Min Temp (°C)': 'mean', 
                                                            'Mean Temp (°C)': 'mean',
                                                            'Total Rain (mm)': 'sum',
                                                            'Total Snow (cm)': 'sum',
                                                            'Total Precip (mm)': 'sum'}).reset_index()
                                      
# to rename the columns to be consistent with the initial data
monthly_data_clean.rename(columns={'Max Temp (°C)' : 'Mean Max Temp (°C)',
                             'Min Temp (°C)' : 'Mean Min Temp (°C)'},
                          inplace = True)

# to create the 'Extr Max Temp (°C)' column
extr_max_temp = monthly_data.groupby(['Longitude', 
                                      'Latitude',
                                      'Station Name',
                                      'Climate ID',
                                      'Year',
                                      'Month',
                                      'GeoUID',
                                      'Region.Name',
                                      'provincename'])['Max Temp (°C)'].max().reset_index()

# to rename the columns to be consistent with the initial data
extr_max_temp.rename(columns={'Max Temp (°C)' : 'Extr Max Temp (°C)'},
                    inplace = True)

# to merge the extr_max_temp column with the dataframe
monthly_data_clean = pd.merge(monthly_data_clean, 
                             extr_max_temp, 
                             how='left', 
                             on=['Longitude', 'Latitude', 
                                 'Station Name', 'Climate ID', 
                                 'Year', 'Month', 'GeoUID', 
                                 'Region.Name', 'provincename'])

# to create the 'Extr Min Temp (°C)' column
extr_min_temp = monthly_data.groupby(['Longitude', 
                                      'Latitude',
                                      'Station Name',
                                      'Climate ID',
                                      'Year',
                                      'Month',
                                      'GeoUID',
                                      'Region.Name',
                                      'provincename'])['Min Temp (°C)'].min().reset_index()
                               
# to rename the columns to be consistent with the initial data
extr_min_temp.rename(columns={'Min Temp (°C)' : 'Extr Min Temp (°C)'},
                    inplace = True)

# to merge the extr_min_temp column with the dataframe
monthly_data_clean = pd.merge(monthly_data_clean, 
                             extr_min_temp, 
                             how='left', 
                             on=['Longitude', 'Latitude', 
                                 'Station Name', 'Climate ID', 
                                 'Year', 'Month', 'GeoUID', 
                                 'Region.Name', 'provincename'])

# convert column to date time format
monthly_data['Date/Time'] = pd.to_datetime(monthly_data['Date/Time'])

# Find the row with the maximum day for each group
last_day = monthly_data.loc[monthly_data.groupby(['Longitude', 'Latitude', 
                                                  'Station Name', 'Climate ID', 
                                                  'Year', 'Month', 'GeoUID', 
                                                  'Region.Name', 'provincename'])['Day'].idxmax()]
  
last_day = last_day[['Longitude', 'Latitude', 'Station Name', 
                      'Climate ID', 'Year', 'Month', 'GeoUID', 
                      'Region.Name', 'provincename', 'Snow on Grnd (cm)']].copy()

# to rename the columns to be consistent with the initial data
last_day.rename(columns={'Snow on Grnd (cm)' : 'Snow Grnd Last Day (cm)'},
                    inplace = True)
                    
## to merge the extr_min_temp column with the dataframe
monthly_data_clean = pd.merge(monthly_data_clean, 
                             last_day, 
                             how='left', 
                             on=['Longitude', 'Latitude', 
                                 'Station Name', 'Climate ID', 
                                 'Year', 'Month', 'GeoUID', 
                                 'Region.Name', 'provincename'])                     

# to replace null values with 0
monthly_data.fillna(0, inplace=True)

monthly_data.isnull().sum()

# convert to numeric
monthly_data['Spd of Max Gust (km/h)'] = pd.to_numeric(monthly_data['Spd of Max Gust (km/h)'], errors='coerce')

# create spd max gust by getting the max from the same column
max_speed_index = monthly_data.groupby(['Longitude', 'Latitude', 'Station Name', 'Climate ID', 'Year', 'Month', 'GeoUID', 'Region.Name', 'provincename'])['Spd of Max Gust (km/h)'].idxmax()

# to replace null values with 0 max_speed_index
max_speed_index.fillna(0, inplace=True)

spd_max_gust = monthly_data.loc[max_speed_index, ['Longitude', 'Latitude', 'Station Name', 'Climate ID', 'Year', 'Month', 'GeoUID', 'Region.Name', 'provincename', 'Spd of Max Gust (km/h)', 'Dir of Max Gust (10s deg)']]

                   
## to merge the extr_min_temp column with the dataframe
monthly_data_clean = pd.merge(monthly_data_clean, 
                             spd_max_gust, 
                             how='left', 
                             on=['Longitude', 'Latitude', 
                                 'Station Name', 'Climate ID', 
                                 'Year', 'Month', 'GeoUID', 
                                 'Region.Name', 'provincename'])

monthly_data_clean.tail().T

max_speed_index.isnull().sum()

# to check that it's for 1 province only
monthly_data_clean['provincename'].unique()

# to check the count of rows of the original data and the clean data
print(weather_clean.shape)
print(monthly_data_clean.shape)

# to convert to string
monthly_data_clean['GeoUID'] = monthly_data_clean['GeoUID'].astype('int').astype('object')
# to convert to datetime format
monthly_data_clean['Date/Time'] = pd.to_datetime(monthly_data_clean[['Year', 'Month']].assign(DAY=1))
# to check dataset
monthly_data_clean.info()

# to arrange columns
monthly_data_clean = monthly_data_clean[['Longitude', 'Latitude', 'Station Name', 
                                         'Climate ID', 'Date/Time', 'Year', 'Month', 
                                         'Mean Max Temp (°C)', 'Mean Min Temp (°C)',
                                         'Mean Temp (°C)','Extr Max Temp (°C)',
                                         'Extr Min Temp (°C)', 'Total Rain (mm)',
                                         'Total Snow (cm)', 'Total Precip (mm)',
                                         'Snow Grnd Last Day (cm)', 'Dir of Max Gust (10s deg)',
                                         'Spd of Max Gust (km/h)','GeoUID', 'Region.Name', 'provincename']]

# to check date range
print(monthly_data_clean['Date/Time'].min())
print(monthly_data_clean['Date/Time'].max()) # to remove data after Jan 2024 since there are incomplete

monthly_data_clean = monthly_data_clean[monthly_data_clean['Date/Time'] <= '2024-01-01']

# to check date range; confirm that only unti; Jan 2024
print(monthly_data_clean['Date/Time'].min())
print(monthly_data_clean['Date/Time'].max())

# to output weather data 2007 to 2017 into csv format

csv_path = 'weather_geography_merged_2007_to_2017_SK.csv'
monthly_data_clean.to_csv(csv_path, 
                   index = False)

# to check converted data if consistent with the expected values (canada site)
weather_2024 = monthly_data_clean[monthly_data_clean['Date/Time'] >= '2024-01-01']
weather_2024.head().T

# to extract the station name and climate ID for each provincename; for further weather data extraction

downloaded_SK = monthly_data_clean[['provincename', 'Station Name', 'Climate ID']].drop_duplicates()
downloaded_SK_df = pd.DataFrame(downloaded_SK, columns = ['provincename', 'Station Name', 'Climate ID'])

csv_path = 'downloaded_SK.csv'
downloaded_SK_df.to_csv(csv_path,
                            index = False)
