==> Part II/03_Weather_2018 to 2024_Merged.ipynb <==
## WEATHER (2018 - 2024) FOR 5 PROVINCES ##

import pandas as pd

# load data

AB = pd.read_csv('C:/Users/chust/OneDrive/Documents/GitHub/Climate-Change-and-its-Economic-Impact-DAB422-24W-004-Group9/Part II/weather_geography_merged_2018_to_2024_AB.csv')
BC = pd.read_csv('C:/Users/chust/OneDrive/Documents/GitHub/Climate-Change-and-its-Economic-Impact-DAB422-24W-004-Group9/Part II/weather_geography_merged_2018_to_2024_BC.csv')
ON = pd.read_csv('C:/Users/chust/OneDrive/Documents/GitHub/Climate-Change-and-its-Economic-Impact-DAB422-24W-004-Group9/Part II/weather_geography_merged_2018_to_2024_ON.csv')
QC = pd.read_csv('C:/Users/chust/OneDrive/Documents/GitHub/Climate-Change-and-its-Economic-Impact-DAB422-24W-004-Group9/Part II/weather_geography_merged_2018_to_2024_QC.csv')
SK_1 = pd.read_csv('C:/Users/chust/OneDrive/Documents/GitHub/Climate-Change-and-its-Economic-Impact-DAB422-24W-004-Group9/Part II/weather_geography_merged_2018_to_2024_SK.csv')
SK_2 = pd.read_csv('C:/Users/chust/OneDrive/Documents/GitHub/Climate-Change-and-its-Economic-Impact-DAB422-24W-004-Group9/Part II/weather_geography_merged_2007_to_2017_SK.csv')

AB

BC

ON

QC

SK_1

print(SK_1['Date/Time'].min())
print(SK_1['Date/Time'].max())

SK_2

print(SK_2['Date/Time'].min())
print(SK_2['Date/Time'].max())

final_merged_weather = pd.concat([AB, BC, ON, QC, SK_1, SK_2], axis = 0)

final_merged_weather

csv_path = 'final_merged_weather_2018_2024.csv'
final_merged_weather.to_csv(csv_path,
                            index= False)
