==> Part II/04_Productivity.ipynb <==
## PRODUCTIVITY FOR 5 PROVINCES ##

# import libraries

import pandas as pd

# ignore all warnings

import warnings
warnings.filterwarnings("ignore")

# to load productivity data

productivity = pd.read_csv('combined_productivity_clean.csv') # 

productivity.head()

productivity.info()

#checking for missing values
productivity.isnull().sum()

# update data column to datatime format
productivity['Date/Time'] = pd.to_datetime(productivity['Date'])
productivity

# to check  productivity date range
print(productivity['Date/Time'].min())
print(productivity['Date/Time'].max())

# to remove 1997 data to be consistent with the weather dataset
prod_clean = productivity[productivity['Date/Time'] >= '1998-01-01']
print(prod_clean['Date/Time'].min())
print(prod_clean['Date/Time'].max())

# to filter 5 provinces
filter = ['Quebec', 'Ontario', 'British Columbia', 'Alberta', 'Saskatchewan']
prod_clean = prod_clean[prod_clean['provincename'].isin(filter)]
print(prod_clean['provincename'].unique())
print(prod_clean.info())

# to load weather_geography_merged data
weather_1998_2018Feb = pd.read_csv('weather_geography_merged_1998_to_2018Feb.csv')
weather_2018_2024 = pd.read_csv('final_merged_weather_2018_2024.csv')
SK_2007_2017 = pd.read_csv('weather_geography_merged_2007_to_2017_SK.csv')

print(weather_2018_2024[weather_2018_2024['provincename'] == 'Quebec']['Date/Time'].min())
print(weather_2018_2024[weather_2018_2024['provincename'] == 'Quebec']['Date/Time'].max())

# to convert the "Date/Time" column to date format
weather_1998_2018Feb['Date/Time'] = pd.to_datetime(weather_1998_2018Feb['Date/Time'])
weather_2018_2024['Date/Time'] = pd.to_datetime(weather_2018_2024['Date/Time'])
SK_2007_2017['Date/Time'] = pd.to_datetime(SK_2007_2017['Date/Time'])

# to remove overlapping data (2018 Jan and Feb data)
weather_1998_2017 = weather_1998_2018Feb[weather_1998_2018Feb['Date/Time'] <= '2017-12-01']

# # to filter the 5 provinces
filter = ['Quebec', 'Ontario', 'British Columbia', 'Alberta', 'Saskatchewan']
weather_1998_2017 = weather_1998_2017[weather_1998_2017['provincename'].isin(filter)]
print(weather_1998_2017['provincename'].unique())

# # to remove data from Feb 2024
weather_2018_2024 = weather_2018_2024[weather_2018_2024['Date/Time'] <= '2024-01-01']

print(weather_1998_2017['Date/Time'].min())
print(weather_1998_2017['Date/Time'].max())

print(weather_1998_2017[weather_1998_2017['provincename'] == 'Saskatchewan']['Date/Time'].min())
print(weather_1998_2017[weather_1998_2017['provincename'] == 'Saskatchewan']['Date/Time'].max())

print(SK_2007_2017['Date/Time'].min())
print(SK_2007_2017['Date/Time'].max())

SK_2007_2017 = SK_2007_2017[SK_2007_2017['Date/Time'] > '2007-11-01']

print(SK_2007_2017['Date/Time'].min())
print(SK_2007_2017['Date/Time'].max())

# to check date range of productivity

print(prod_clean['Date/Time'].min())
print(prod_clean['Date/Time'].max())

weather_1998_2018Feb.shape

weather_2018_2024.shape

# to merge the 2 weather datasets
merged_weather = pd.concat([weather_1998_2017, weather_2018_2024, SK_2007_2017])
merged_weather.info()

# to check shape of the datasets
print(prod_clean.shape)
print(merged_weather.shape)

# to merge data:

merged_data = pd.merge(prod_clean, 
                       merged_weather,
                       on = ['GeoUID', 'Date/Time', 'provincename'],
                       how = 'right'
)
###### UPDATED: TO REVERT BACK IF DOESN'T WORK!!! how: from right to outer
    # used 'right' to only get productivity data that have weather data
    # not all GeoUID have weather data; we won't be able to forecast productivity for those GeoUID without the weather data

# to check shape of the datasets
print(prod_clean.shape)
print(merged_weather.shape)
print(merged_data.shape)

print(merged_data[merged_data['provincename'] == 'Quebec']['Date/Time'].min())
print(merged_data[merged_data['provincename'] == 'Quebec']['Date/Time'].max())

print(merged_data[merged_data['provincename'] == 'Quebec']['Date/Time'].dt.year.unique())

# to check date range
print(merged_data['Date/Time'].min())
print(merged_data['Date/Time'].max())

# to remove unnecessary columns

merged_data.drop(columns = ['Date', 'Year', 'Month',
                            'census_year_ref', 'Dominant_NAICS',
                            'colourval'],
                 inplace = True)

# to arrange columns

columns_to_move = ['provincename', 'Longitude', 'Latitude', 'Region.Name', 'Date/Time']
remaining_columns = [col for col in merged_data.columns if col not in columns_to_move]
new_order = columns_to_move + remaining_columns
merged_data = merged_data[new_order]

# remove flag columns
flag_columns = [col for col in merged_data.columns if 'Flag' in col]
merged_data = merged_data.drop(columns = flag_columns)

# to remove stations that does not have productivity data
filter = ['KEMANO',
    'MCINNES ISLAND',
    'QUATSINO',
    'QUATSINO LIGHTSTATION',
    'SARTINE ISLAND (AUT)',
    'CHALK RIVER AECL',
    'OTTAWA CDA',
    "L'ETAPE"]
merged_data = merged_data[~merged_data['Station Name'].isin(filter)]
# print(merged_data['provincename'].unique())
print(merged_data.info())

merged_data['Snow Grnd Last Day (cm)'].fillna(0, inplace = True)

merged_data.isnull().sum()

# # columns
# columns_list = [
#     'provincename',
#     'Longitude',
#     'Latitude',
#     'Region.Name',
#     'Date/Time',
#     'Population',
#     'GeoUID',
#     'Station Name',
#     'Climate ID',
#     'Mean Max Temp (°C)',
#     'Mean Min Temp (°C)',
#     'Mean Temp (°C)',
#     'Extr Max Temp (°C)',
#     'Extr Min Temp (°C)',
#     'Total Rain (mm)',
#     'Total Snow (cm)',
#     'Total Precip (mm)',
#     'Snow Grnd Last Day (cm)',
#     'Dir of Max Gust (10s deg)',
#     'Spd of Max Gust (km/h)'
# ]

# #melting the DataFrame to convert "Productivity areas" columns into rows
# merged_data = pd.melt(merged_data, id_vars=columns_list, var_name='productivity_sector', value_name='GDP_Value')

# #checking changes
# merged_data

#Columns with missing values
columns_with_missing_values = ['Mean Max Temp (°C)', 'Mean Min Temp (°C)', 'Mean Temp (°C)', 'Extr Max Temp (°C)', 'Extr Min Temp (°C)']

#Create a 'Month' column from 'Date/Time'
merged_data['Month'] = merged_data['Date/Time'].dt.month

#Iterate through each column and fill missing values with the mean of the corresponding months
for column in columns_with_missing_values:
    # Create a mask for missing values in the current column
    missing_values_mask = merged_data[column].isnull()
    
    # Calculate the mean for each month in the column
    monthly_means = merged_data.groupby('Month')[column].mean()
    
    # Fill missing values with the mean of the corresponding month
    merged_data.loc[missing_values_mask, column] = merged_data.loc[missing_values_mask, 'Month'].map(monthly_means)

# Drop the 'Month' column if you don't want to keep it
merged_data.drop(columns=['Month'], inplace=True)

# to ouput data

csv_path = 'merged_clean_data.csv'
merged_data.to_csv(csv_path,
                   index = False)

# this data contains the geojson, weather, and productivity
